Finished Playgame Method
Playgame Method words correctly. Several questions about properly setting the door class. Is there a way to set up the door class so that the user only needs to set whether or not the user will switch? The other three elements of the class are all overwritten Playgame method anyway. it would make more sense for these to start as NULL.